refactor(ai): estrai il prompt document synthesis
